requirements.txt
Flask
python-dotenv
uv